moved xbmcclient lib file into lib/xbmc dir
